fix(chat): fix monaco editor empty folder behaviour (Issue #6222, #6223)
