+++ apps/chat/src/store/codeEditor/codeEditor.types.ts
@@ -9,4 +9,5 @@ export interface CodeEditorState {
   }[];
   fileContentLoadingStatus: UploadStatus;
   selectedFileId: string | undefined;
+  sourcesFolderId: string | undefined;
 }
